[PKTGEN] : fix an oops when used with bonding driver (Tien ChenLi)
Back-port from Tien ChenLi's 2.6 patch described this below.

"I fixed a bug in pktgen so it won't cause oops when used with
balance-tlb or balance-alb bonding driver".

The root cause is that the bond_alb_xmit in bonding will peek the
destination address in packet via the skb->nh.iph pointer, generally
this will be filled by upper layer network driver, but the packet
generated by pktgen will be sent to device driver so it will need to
set this pointer correctly. The other two pointers are not necessary
for now, they are set to avoid similar problem.