[GCC4] add preliminary support for GCC 4 (Mikael Pettersson)
This is the well tested part of Mikael Pettersson's GCC4 patches.
It offers support for GCC4 on x86, x86_64 and ppc. Other archs
untested. The risk of regression on older compilers at this stage
should be close to zero. At least, x86 with allmodconfig still builds
with gcc-2.95.3.

Other places have been carefully fixed but the code has not been run
yet. They will be merged individually in further patches.