[PATCH] i386: remove unsigned long long cast in __pte() macro.
From PaX Team :
"I've just seen the commit of the __pte() macro fixes and i'm not
 sure that a blind unsigned long long cast is the proper way to
 handle the warnings. the thing is, if there's a warning as Ralf
 said, then it means that the user of the macro is not passing the
 proper argument to the macro, hence there's a likelyhood that the
 caller itself may not be doing what it intended to do."

"I don't know about MIPS, but under i386/PAE the higher 32 bits do
 carry 'interesting' bits (not only physical page number bits but
 on NX capable CPUs the NX bit itself as well), so they can't be
 blindly made 0 by doing an unsigned long long cast, it should be
 decided by the user of the macro instead."

"So my suggestion would be to fix the callers instead of doing the
 cast, that is, have gcc point out everyone not passing a properly
 set up argument, figure out if a cast is proper (as i said, at
 least on PAE it will loose important bits) and fix the caller."

Also, the only user seems to be drm-40/ffb_drv.c, which is only
enabled for sparc64. So the warning would never appear anyway.

Acked-By: Willy Tarreau <w@1wt.eu>