[PATCH] restore ability to set disk geometry on command line
The second patch is a two-liner, which adds (restores?) the possibility to
specify your own disk geometry in the kernel command line. Default values
of 255 heads and 63 sectors can be nonoptimal, i.e. a small part of the
disk remains unused.

BTW sometimes it is desirable to set a specific geometry on small devices
such as flash disks.