[PATCH] lvm: update to latest fixes from the LVM package
greg@enjellic.com suggested to apply the recommended patches from
the LVM package, as every LVM user will have to generate them and
apply them anyway. Heinz Mauelshagen agreed, except for the list_del
chunk in the snapshot code which was identified as responsible for
oopses encountered in RHEL3.

Now with this patch in 2.4, it should be safe(r) to not apply the
LVM patches anymore.