[PATCH] notsc support for x86_64
This patch deals with a problem with unsynced TSCs on dual core Athlon 64.
The time goes sometimes backwards with TSC based gettimeofday(), while the
dummy routine gives time resolution of 10 milliseconds, which is too coarse
for some applications. There are no HPET timers on the mainboard, so I ported
the good old gettimeoffset_slow routine from arch/i386. It seems to work and
gives nice time resolution of about 5 microseconds.

In order to use this code, one will have to specify the "notsc" option on
the kernel command line, just as on x86.

Kernel 2.6 does not need this code as it uses different techniques to
workaround the unsynced TSCs.